requirements.txt
# Core dependencies
requests>=2.28.0
PyMySQL>=1.0.2
DBUtils>=3.0.0
SQLAlchemy>=1.4.0
redis>=4.3.0
pydantic>=1.10.0

# Crypto dependencies
cryptography>=3.4.8
pycryptodome>=3.15.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Optional dependencies for web API
fastapi>=0.85.0
uvicorn>=0.18.0

# Async support
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
celery>=5.2.0
redis>=4.3.0  # Already included above, used as Celery broker
flower>=1.2.0  # Celery monitoring
kombu>=5.2.0  # Celery messaging

# Configuration management
python-dotenv>=0.19.0
pyyaml>=6.0

# Monitoring and metrics
prometheus-client>=0.15.0
psutil>=5.9.0